Fix server request sent delay to be non-negative

+++ pinot-core/src/main/java/org/apache/pinot/core/transport/AsyncQueryResponse.java
@@ -128,7 +128,7 @@ public class AsyncQueryResponse implements QueryResponse {
     _responseMap.get(serverRoutingInstance).markRequestSubmitted();
   }
 
-  void markRequestSent(ServerRoutingInstance serverRoutingInstance, long requestSentLatencyMs) {
+  void markRequestSent(ServerRoutingInstance serverRoutingInstance, int requestSentLatencyMs) {
     _responseMap.get(serverRoutingInstance).markRequestSent(requestSentLatencyMs);
   }
